DO NOT MERGE Revert to libphonenumber v7.0.11
This change reverts to the previously-submitted v7.0.11 version of
libphonenumber. It removes the dependency on protobufs thereby saving
disk / memory.

This is effectively a revert of commits
6b489f77ae70bea07f3a8fbd877368b4e0f3142c and
3bd754c0967ed003578e460faa83ad8e777a636b, and keeping the subsequent
changes.

Some additional edits have been made to build.gradle and
build-full.gradle to keep the gradle build working and remove the
dependency on protos.

Bug: 28341523
Change-Id: Ic35f775dd11095c31625d56492916da783abf36d